Add DOM manipulation and accessibility changes
